Update manager to receive first input
