+++ app/app.py
@@ -1,24 +1,12 @@
-from flask import Flask, request, render_template, jsonify
-from flask_cors import CORS, cross_origin
-from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import Column, ForeignKey, Integer, String, PickleType
-import json
-from jsonschema import validate, ValidationError, SchemaError
-from pathlib import Path
+from flask import Flask
+from flask_cors import CORS
+from views import optimization_blueprint
+from config import DATABASE_URL
+from models import Base
+from db import engine
 # Import of the models
 # https://www.compose.com/articles/using-postgresql-through-sqlalchemy/
 
-# Schema that has to be validated against the uploaded json
-JSON_SCHEMA_UPLOAD = "./json_schema/schema_upload.json"
-
-# Folder for optimization data
-OPTIMIZATION_DATA = "/optimization-data/"
-
-# File ending for json
-JSON_ENDING = ".json"
-
-# Database to our optimization postgres
-DATABASE_URL = "postgresql+psycopg2://root:root@postgres:5432/optimization"
 
 # Create a flask app
 app = Flask(__name__)
@@ -27,147 +15,11 @@ app.config["DEBUG"] = True
 app.config["SQLALCHEMY_DATABASE_URI"] = DATABASE_URL
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 # Enable cross origin resource sharing
+app.register_blueprint(optimization_blueprint)
+
 CORS(app)
 
-# Add the created app to our database
-# Basically our empty database with its models gets the information to which existing database to connect
-db = SQLAlchemy(app)
-
-
-class OptimizationTask(db.Model):
-    __tablename__ = "optimization_tasks"
-
-    author = Column(String)
-    # Column for project
-    project = Column(String)
-    # # Column for calculation_id
-    # calculation_id = db.Column(db.String)
-    # # Column for model_id
-    # model_id = db.Column(db.String)
-    # Column for optimization_id
-    optimization_id = Column(String, primary_key=True)
-    # Column for type
-    type = Column(String)
-    # Column for version
-    # version = db.Column(db.String)
-    # Column for optimization
-    optimization = Column(PickleType)
-    # Column for data
-    # data = db.Column(db.PickleType)
-
-    def __init__(self, author, project, optimization_id, opt_type, optimization, **args):
-        super().__init__(**args)
-
-        self.author = author
-        self.project = project
-        # self.calculation_id = calculation_id
-        # self.model_id = model_id
-        self.optimization_id = optimization_id
-        self.type = opt_type
-        # self.version = version
-        self.optimization = optimization
-        # self.data = data
-
-
-# Now we want to create all our tables/models (here optimization_tasks table)
-db.create_all()
-
-# Our main upload page where we put our json
-@app.route("/upload", methods=['GET', 'POST'])
-@cross_origin()
-def upload_file():
-    # This is our most important path
-    if request.method == 'POST':
-        # https://stackoverflow.com/questions/46136478/flask-upload-how-to-get-file-name
-        # See if there's a file in our selection field
-        if not request.files.get('file', None):
-            error = "No file selected!"
-            return render_template('upload.html', error=error)
-
-        # Get json from request
-        file_upload = request.files["file"]
-        req_data = json.load(file_upload)
-
-        # Load validation schema form json_schema
-        with open(JSON_SCHEMA_UPLOAD, "r") as f:
-            schema_upload = json.load(f)
-
-        # Check the json against our json schema as provided in ./json_schema/schema_upload.json
-        try:
-            # Validate file
-            validate(instance=req_data, schema=schema_upload)
-        # In case there's an error with our data
-        except ValidationError as e:
-            # A message to show where there's a problem in the json file
-            error = {
-                "message": f"Validation failed. {e.message}",
-                "code": e.validator,
-                "schemapath": e.schema_path
-            }
-
-            return render_template('upload.html', error=error)
-        # In case there's an error with our schema itself
-        except SchemaError as e:
-            # Print error
-            error = str(e)
-
-            return render_template('upload.html', error=error)
-
-        # Now since no errors appeared the task will be added to the database
-        # First we split our request data into our columns
-
-        # Author, project are not necessary and will be replaced by standards if not given
-        author = req_data.get("author", "Max Mustermann")
-        project = req_data.get("project", "Standardprojekt")
-        # optimization_id, optimization and type are necessary and can be retrieved directly
-        optimization_id = req_data["optimization_id"]
-        opt_type = req_data["type"]
-        optimization = req_data["optimization"]
-        # data is also necessary and can be retrieved directly; data will be written as file, as it's a big chunk
-        # of data and can easily be stored/loaded as json
-        data = req_data["data"]
-
-        # Create instance of task
-        optimizationtask = OptimizationTask(
-                                author=author,
-                                project=project,
-                                optimization_id=optimization_id,
-                                opt_type=opt_type,
-                                optimization=optimization
-                            )
-
-        # Where to store the data
-        filepath = f"{OPTIMIZATION_DATA}{optimization_id}{JSON_ENDING}"
-
-         # Try adding file to folder and adding job to table
-        try:
-            # Open created filepath
-            with open(filepath, "w") as f:
-                # Write json to it
-                json.dump(data, f)
-
-            # Add task to table
-            db.session.add(optimizationtask)
-            # Push it to the server
-            db.session.commit()
-        except Exception:
-            Path(filepath).unlink()
-
-        return jsonify(req_data)
-
-    if request.method == 'GET':
-        if request.content_type == "application/json":
-            return json.dumps({
-                'message': "test"
-            })
-        return render_template('upload.html')
-
-
-# Our optimization page where we can see the progress of running optimizations
-@app.route("/optimization", methods=["GET"])
-@cross_origin()
-def show_optimizations():
-    pass
+Base.metadata.create_all(engine)
 
 
 if __name__ == "__main__":
