+++ app/config.py
@@ -1,0 +1,8 @@
+# Database to our optimization postgres
+DATABASE_URL = "postgresql+psycopg2://root:root@postgres:5432/optimization"
+
+# Folder for optimization data
+OPTIMIZATION_DATA = "/optimization-data/"
+
+# File ending for json
+JSON_ENDING = ".json"
