+++ manager/requirements.txt
@@ -1,3 +1,8 @@
+Flask==1.0.2
+Flask-Cors==3.0.7
+Flask-Migrate==2.5.2
+Flask-Script==2.0.6
+Flask-SQLAlchemy==2.4.0
 SQLAlchemy==1.3.7
 psycopg2-binary==2.8.3
 scipy
