+++ worker/worker.py
@@ -1,8 +1,13 @@
+import os.path
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), 'opt_app'))
 from time import sleep
-
+# from app.models import DATABASE_URL
+# from manager.models import CalculationTask
 
 def run():
     while True:
+        # current_calc_task = CalculationTask.query.first()
         print('No jobs, sleeping for 1 minute')
         sleep(60)
         continue
